requirements.txt
streamlit
Python-dotenv
google.generativeai